+++ testing/endtoend/components/BUILD.bazel
@@ -6,7 +6,6 @@ go_library(
     srcs = [
         "beacon_node.go",
         "boot_node.go",
-        "eth1.go",
         "lighthouse_beacon.go",
         "lighthouse_validator.go",
         "log.go",
@@ -27,19 +26,17 @@ go_library(
         "//config/features:go_default_library",
         "//config/params:go_default_library",
         "//contracts/deposit:go_default_library",
-        "//contracts/deposit/mock:go_default_library",
         "//crypto/bls:go_default_library",
         "//encoding/bytesutil:go_default_library",
         "//io/file:go_default_library",
         "//runtime/interop:go_default_library",
+        "//testing/endtoend/components/eth1:go_default_library",
         "//testing/endtoend/helpers:go_default_library",
         "//testing/endtoend/params:go_default_library",
         "//testing/endtoend/types:go_default_library",
         "//testing/util:go_default_library",
         "//validator/keymanager:go_default_library",
         "@com_github_ethereum_go_ethereum//accounts/abi/bind:go_default_library",
-        "@com_github_ethereum_go_ethereum//accounts/keystore:go_default_library",
-        "@com_github_ethereum_go_ethereum//common:go_default_library",
         "@com_github_ethereum_go_ethereum//common/hexutil:go_default_library",
         "@com_github_ethereum_go_ethereum//ethclient:go_default_library",
         "@com_github_ethereum_go_ethereum//rpc:go_default_library",
