+++ testing/endtoend/components/eth1/BUILD.bazel
@@ -1,0 +1,29 @@
+load("@prysm//tools/go:def.bzl", "go_library")
+
+go_library(
+    name = "go_default_library",
+    testonly = True,
+    srcs = [
+        "constants.go",
+        "miner.go",
+        "node.go",
+        "node_set.go",
+    ],
+    importpath = "github.com/prysmaticlabs/prysm/testing/endtoend/components/eth1",
+    visibility = ["//testing/endtoend:__subpackages__"],
+    deps = [
+        "//contracts/deposit/mock:go_default_library",
+        "//io/file:go_default_library",
+        "//testing/endtoend/helpers:go_default_library",
+        "//testing/endtoend/params:go_default_library",
+        "//testing/endtoend/types:go_default_library",
+        "@com_github_ethereum_go_ethereum//accounts/abi/bind:go_default_library",
+        "@com_github_ethereum_go_ethereum//accounts/keystore:go_default_library",
+        "@com_github_ethereum_go_ethereum//ethclient:go_default_library",
+        "@com_github_ethereum_go_ethereum//log:go_default_library",
+        "@com_github_ethereum_go_ethereum//rpc:go_default_library",
+        "@com_github_pkg_errors//:go_default_library",
+        "@com_github_sirupsen_logrus//:go_default_library",
+        "@io_bazel_rules_go//go/tools/bazel:go_default_library",
+    ],
+)
